1. Move CoSignatures to libCrypto. 2. Remove some unused code.
