+++ src/libLookup/Lookup.cpp
@@ -440,45 +440,6 @@ VectorOfNode Lookup::GetSeedNodes() const {
   }
 }
 
-std::once_flag generateReceiverOnce;
-
-Address GenOneReceiver() {
-  static Address receiverAddr;
-  std::call_once(generateReceiverOnce, []() {
-    auto receiver = Schnorr::GenKeyPair();
-    receiverAddr = Account::GetAddressFromPublicKey(receiver.second);
-    LOG_GENERAL(INFO, "Generate testing transaction receiver " << receiverAddr);
-  });
-  return receiverAddr;
-}
-
-Transaction CreateValidTestingTransaction(PrivKey& fromPrivKey,
-                                          PubKey& fromPubKey,
-                                          const Address& toAddr,
-                                          const uint128_t& amount,
-                                          uint64_t prevNonce) {
-  unsigned int version = 0;
-  auto nonce = prevNonce + 1;
-
-  // LOG_GENERAL("fromPrivKey " << fromPrivKey << " / fromPubKey " << fromPubKey
-  // << " / toAddr" << toAddr);
-
-  Transaction txn(version, nonce, toAddr, make_pair(fromPrivKey, fromPubKey),
-                  amount, 1, 1, {}, {});
-
-  // bytes buf;
-  // txn.SerializeWithoutSignature(buf, 0);
-
-  // Signature sig;
-  // Schnorr::Sign(buf, fromPrivKey, fromPubKey, sig);
-
-  // bytes sigBuf;
-  // sig.Serialize(sigBuf, 0);
-  // txn.SetSignature(sigBuf);
-
-  return txn;
-}
-
 bool Lookup::GenTxnToSend(size_t num_txn, vector<Transaction>& shardTxn,
                           vector<Transaction>& DSTxn) {
   vector<Transaction> txns;
