+++ src/libNode/Node.cpp
@@ -1436,38 +1436,6 @@ bool GetOneGenesisAddress(Address& oAddr) {
   return true;
 }
 
-/// generate transation from one to many random accounts
-/*vector<Transaction> GenTransactionBulk(PrivKey& fromPrivKey, PubKey&
-fromPubKey, size_t n)
-{
-    vector<Transaction> txns;
-
-    // FIXME: it's a workaround to use the first genensis account
-    // auto receiver = Schnorr::GenKeyPair();
-    // auto receiverAddr = Account::GetAddressFromPublicKey(receiver.second);
-
-    // alternative 1: use first genesis address
-    // Address addr;
-    // if (not GetOneGenesisAddress(addr))
-    // {
-    // return txns;
-    // }
-    // auto receiverAddr = addr;
-
-    // alternative 2: use a fresh address throughout entire lifetime
-    auto receiverAddr = GenOneReceiver();
-
-    txns.reserve(n);
-    for (auto i = 0u; i != n; i++)
-    {
-        auto txn = CreateValidTestingTransaction(fromPrivKey, fromPubKey,
-                                                 receiverAddr, i);
-        txns.emplace_back(txn);
-    }
-
-    return txns;
-}*/
-
 bool Node::ProcessSubmitMissingTxn(const zbytes& message, unsigned int offset,
                                    [[gnu::unused]] const Peer& from) {
   if (LOOKUP_NODE_MODE) {
